Change license to MIT
PyPi says the license is MIT but this file says it's GPL. This needs to
become consistently declared. I can't use GPL licensed code because how
it forces all code that uses it to be open sourced as well. If you merge
this pull request, we can consistently relicense it as MIT.